requirements.txt
elasticsearch7
python-geoip-python3
python-geoip-geolite2
